requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
gunicorn==21.2.0
